fixed float to int

+++ task_1/main.py
@@ -3,18 +3,18 @@ from pathlib import Path
 def total_salary(path):
     total = 0
     developers_count = 0
-    
+
     try:
         with open(path, mode="r", encoding="utf-8") as file:
             for line in file:
                 _, salary = line.strip().split(",")
-                total += float(salary)
+                total += int(salary)
                 developers_count += 1
 
         if developers_count == 0:
             return 0, 0
                 
-        average_salary = total / developers_count
+        average_salary = int(total / developers_count)
         return total, average_salary
 
     except (FileNotFoundError, ValueError) as e:
